Refactor Echonet Lite integration by consolidating property handling
- Moved property-related classes (BinaryProp, NumericProp, EnumProp) to a new `prop.py` file for better organization.
- Updated imports across various modules to reference the new property classes.
- Simplified runtime management by integrating the RuntimeIssueMonitor and RuntimeController into the `runtime.py` file.
- Removed redundant code and classes from `entity.py` and other modules, enhancing maintainability.
- Adjusted the setup process in `__init__.py` to accommodate the new structure and ensure proper initialization of runtime components.
