+++ custom_components/echonet_lite/cover.py
@@ -24,12 +24,8 @@ from .const import (
     EPC_COVER_POSITION,
 )
 from .coordinator import EchonetLiteCoordinator
-from .entity import (
-    EchonetLiteEntity,
-    EnumProp,
-    NumericProp,
-    setup_echonet_lite_device_platform,
-)
+from .entity import EchonetLiteEntity, setup_echonet_lite_device_platform
+from .prop import EnumProp, NumericProp
 from .types import EchonetLiteConfigEntry
 
 PARALLEL_UPDATES = 1
