+++ custom_components/echonet_lite/light.py
@@ -27,13 +27,8 @@ from .const import (
     EPC_OPERATION_STATUS,
 )
 from .coordinator import EchonetLiteCoordinator
-from .entity import (
-    BinaryProp,
-    EchonetLiteEntity,
-    EnumProp,
-    NumericProp,
-    setup_echonet_lite_device_platform,
-)
+from .entity import EchonetLiteEntity, setup_echonet_lite_device_platform
+from .prop import BinaryProp, EnumProp, NumericProp
 from .types import EchonetLiteConfigEntry
 
 PARALLEL_UPDATES = 1
