+++ custom_components/echonet_lite/lock.py
@@ -16,12 +16,8 @@ from .const import (
     EPC_LOCK_SETTING_2,
 )
 from .coordinator import EchonetLiteCoordinator
-from .entity import (
-    BinaryProp,
-    EchonetLiteEntity,
-    EnumProp,
-    setup_echonet_lite_device_platform,
-)
+from .entity import EchonetLiteEntity, setup_echonet_lite_device_platform
+from .prop import BinaryProp, EnumProp
 from .types import EchonetLiteConfigEntry
 
 PARALLEL_UPDATES = 1
