+++ custom_components/echonet_lite/number.py
@@ -23,9 +23,9 @@ from .coordinator import EchonetLiteCoordinator
 from .entity import (
     EchonetLiteDescribedEntity,
     EchonetLiteEntityDescription,
-    NumericProp,
     setup_echonet_lite_platform,
 )
+from .prop import NumericProp
 from .types import EchonetLiteConfigEntry
 
 PARALLEL_UPDATES = 1
